+++ src/backend/utils/misc/guc.c
@@ -11695,7 +11695,7 @@ show_unix_socket_permissions(void)
 {
 	static char buf[12];
 
-	snprintf(buf, sizeof(buf), "%04o", Unix_socket_permissions);
+	snprintf(buf, sizeof(buf), "%04o", (unsigned) Unix_socket_permissions);
 	return buf;
 }
 
@@ -11704,7 +11704,7 @@ show_log_file_mode(void)
 {
 	static char buf[12];
 
-	snprintf(buf, sizeof(buf), "%04o", Log_file_mode);
+	snprintf(buf, sizeof(buf), "%04o", (unsigned) Log_file_mode);
 	return buf;
 }
 
@@ -11713,7 +11713,7 @@ show_data_directory_mode(void)
 {
 	static char buf[12];
 
-	snprintf(buf, sizeof(buf), "%04o", data_directory_mode);
+	snprintf(buf, sizeof(buf), "%04o", (unsigned) data_directory_mode);
 	return buf;
 }
